+++ packages/client/src/internal/conversions/modernToLegacyWhereClause.ts
@@ -72,6 +72,37 @@ export function modernToLegacyWhereClause<
   objectOrInterface: T,
   rdpNames?: Set<string>,
 ): SearchJsonQueryV2 {
+  const parts = Object.entries(whereClause).map(([key, value]) => ({
+    [key]: value,
+  })) as WhereClause<T, RDPs>[];
+  if (parts.length === 1) {
+    return modernToLegacyWhereClauseInner(
+      whereClause,
+      objectOrInterface,
+      rdpNames,
+    );
+  }
+  return {
+    type: "and",
+    value: parts.map<SearchJsonQueryV2>(
+      v => modernToLegacyWhereClauseInner(v, objectOrInterface, rdpNames),
+    ),
+  };
+}
+
+/** @internal */
+export function modernToLegacyWhereClauseInner<
+  T extends ObjectOrInterfaceDefinition,
+  RDPs extends Record<string, SimplePropertyDef> = {},
+>(
+  whereClause: WhereClause<T, RDPs>,
+  objectOrInterface: T,
+  rdpNames?: Set<string>,
+): SearchJsonQueryV2 {
+  const parts = Object.entries(whereClause);
+
+  invariant(parts.length === 1, "Invalid where clause provided.");
+
   if (isAndClause(whereClause)) {
     return {
       type: "and",
@@ -99,18 +130,7 @@ export function modernToLegacyWhereClause<
     };
   }
 
-  const parts = Object.entries(whereClause);
-
-  if (parts.length === 1) {
-    return handleWherePair(parts[0], objectOrInterface, undefined, rdpNames);
-  }
-
-  return {
-    type: "and",
-    value: parts.map<SearchJsonQueryV2>(
-      v => handleWherePair(v, objectOrInterface, undefined, rdpNames),
-    ),
-  };
+  return handleWherePair(parts[0], objectOrInterface, undefined, rdpNames);
 }
 
 function handleWherePair(
@@ -146,11 +166,16 @@ function handleWherePair(
     ? fullyQualifyPropName(fieldName, objectOrInterface)
     : undefined;
 
+  invariant(
+    field == null
+      || propertyIdentifier == null && (field != null || isRdp != null),
+    "Encountered error constructing where clause: field and propertyIdentifier cannot both be defined",
+  );
+
   if (
     typeof filter === "string" || typeof filter === "number"
     || typeof filter === "boolean"
   ) {
-    propertyIdentifier;
     return {
       type: "eq",
       ...(propertyIdentifier != null