+++ packages/client/src/observable/internal/list/ListQuery.ts
@@ -15,7 +15,6 @@
  */
 
 import type {
-  Logger,
   ObjectOrInterfaceDefinition,
   ObjectSet,
   ObjectTypeDefinition,
@@ -364,63 +363,7 @@ export abstract class ListQuery extends BaseListQuery<
   ): ExtractRelevantObjectsResult;
 
   registerStreamUpdates(sub: Subscription): void {
-    const logger = process.env.NODE_ENV !== "production"
-      ? this.logger?.child({ methodName: "registerStreamUpdates" })
-      : this.logger;
-
-    if (process.env.NODE_ENV !== "production") {
-      logger?.child({ methodName: "observeList" }).info(
-        "Subscribing from websocket",
-      );
-    }
-
-    // FIXME: We should only do this once. If we already have one we should probably
-    // just reuse it.
-
-    const websocketSubscription = this.#objectSet.subscribe({
-      onChange: this.onOswChange.bind(this),
-      onError: this.onOswError.bind(this),
-      onOutOfDate: this.onOswOutOfDate.bind(this),
-      onSuccessfulSubscription: this.onOswSuccessfulSubscription.bind(this),
-    });
-
-    sub.add(() => {
-      if (process.env.NODE_ENV !== "production") {
-        logger?.child({ methodName: "observeList" }).info(
-          "Unsubscribing from websocket",
-        );
-      }
-
-      websocketSubscription.unsubscribe();
-    });
-  }
-
-  protected onOswSuccessfulSubscription(): void {
-    if (process.env.NODE_ENV !== "production") {
-      this.logger?.child(
-        { methodName: "onSuccessfulSubscription" },
-      ).debug("");
-    }
-  }
-
-  protected onOswOutOfDate(): void {
-    if (process.env.NODE_ENV !== "production") {
-      this.logger?.child(
-        { methodName: "onOutOfDate" },
-      ).debug("");
-    }
-  }
-
-  protected onOswError(errors: {
-    subscriptionClosed: boolean;
-    error: any;
-  }): void {
-    if (this.logger) {
-      this.logger?.child({ methodName: "onError" }).error(
-        "subscription errors",
-        errors,
-      );
-    }
+    this.createWebsocketSubscription(this.#objectSet, sub, "observeList");
   }
 
   protected onOswChange(
@@ -451,14 +394,16 @@ export abstract class ListQuery extends BaseListQuery<
         );
       });
     } else if (state === "REMOVED") {
-      this.onOswRemoved(objOrIface, logger);
+      this.onOswRemoved(objOrIface);
     }
   }
 
   protected onOswRemoved(
     objOrIface: Osdk.Instance<ObjectTypeDefinition, never, string, {}>,
-    logger: Logger | undefined,
   ): void {
+    const logger = process.env.NODE_ENV !== "production"
+      ? this.logger?.child({ methodName: "onOswRemoved" })
+      : this.logger;
     this.store.batch({}, (batch) => {
       // Read the truth layer (since not optimistic)
       const existing = batch.read(this.cacheKey);