+++ packages/e2e.generated.catchall/src/generatedNoCheck/index.ts
@@ -71,6 +71,7 @@ export {
   getBusId,
   getBusInterface,
   getBusInterfaceSet,
+  getFirstSecondElementOf2DArray,
   getNamesOfCustomersFromCountry,
   getTodoCount,
   queryTakesAllParameterTypes,