+++ packages/generator-converters.ontologyir/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @osdk/generator-converters.ontologyir
 
+## 2.6.0-beta.6
+
+### Patch Changes
+
+- @osdk/client.unstable@2.6.0-beta.6
+
+## 2.6.0-beta.5
+
+### Patch Changes
+
+- @osdk/client.unstable@2.6.0-beta.5
+
 ## 2.6.0-beta.4
 
 ### Patch Changes